requirements.txt
openai
fastapi
pydantic-settings